Move Check from components to Pages
